[Refactor] Move shadcn components and update imports
