+++ src/components/layouts/Faq/InfoContainer.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { motion } from "framer-motion";
-import { Button } from "@/components/ui/button";
+import { Button } from "@/components/ui/shadcn/button";
 import { ListTodo } from "lucide-react";
 import { Link } from "react-router-dom";
