+++ src/components/layouts/Ranking/Children/Global/GlobalTable/SearchPlayer.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { Input } from "@/components/ui/input";
+import { Input } from "@/components/ui/shadcn/input";
 import { Search } from "lucide-react";
 
 interface Props {
