+++ src/components/layouts/Servers/Fetching/Filtering.tsx
@@ -1,5 +1,5 @@
 import React, { useState, useMemo, useEffect } from "react";
-import { Input } from "@/components/ui/input";
+import { Input } from "@/components/ui/shadcn/input";
 import { Search } from "lucide-react";
 
 export interface Server {
