+++ src/components/ui/shadcn/pagination.tsx
@@ -2,7 +2,7 @@ import * as React from "react";
 import { ChevronLeft, ChevronRight, MoreHorizontal } from "lucide-react";
 
 import { cn } from "@/lib/utils";
-import { ButtonProps, buttonVariants } from "@/components/ui/button";
+import { ButtonProps, buttonVariants } from "@/components/ui/shadcn/button";
 
 const Pagination = ({ className, ...props }: React.ComponentProps<"nav">) => (
     <nav
